Can now switch between fractal types while program is running
