+++ FractalPrototype/src/main/java/com/example/fractalprototype/FractalController.java
@@ -1,19 +1,27 @@
 package com.example.fractalprototype;
 
-import javafx.fxml.FXML;
-import javafx.scene.control.Label;
+import javafx.beans.value.ChangeListener;
+import javafx.beans.value.ObservableValue;
+import javafx.scene.Scene;
 import javafx.scene.layout.Region;
-import javafx.util.Builder;
+import javafx.stage.Stage;
 
 public class FractalController {
-    private final FractalViewBuilder viewBuilder;
-    private FractalModel model = new FractalModel();
+    private final FractalViewBuilder fractalView;
+    private final FractalModel model = new FractalModel();
 
     public FractalController() {
-        viewBuilder = new FractalViewBuilder(model);
+        fractalView = new FractalViewBuilder(model);
+//        model.fractalTypeProperty().addListener(new ChangeListener<String>() {
+//            @Override
+//            public void changed(ObservableValue<? extends String> observableValue,
+//                                String oldValue, String newValue) {
+//                getView();
+//            }
+//        });
     }
 
     public Region getView() {
-        return viewBuilder.build();
+        return fractalView.build();
     }
 }
